Move SupportedCountrySelect to shared components folder.
This is to be used in setup ads.
